refactor: move icons to folder under components
