+++ src/components/Footer/index.tsx
@@ -2,7 +2,7 @@ import './index.scss';
 
 import React from 'react';
 import { Button, Typography } from '@material-ui/core';
-import { CopyIcon } from '../../icons';
+import { CopyIcon } from '../icons';
 import { copyToClipboard } from '../../utilities';
 
 interface FooterProps {
